ENH: Execute FSL FAST only with ``--fs-no-reconall``
This PR restructures the anatomical workflow so that the ``recon-all``
based graph does not execute FSL FAST, and rather projects the
``aseg`` segmentation to a 3-tissue brain segmentation with labels
according to BIDS.

This should provide the confounds workflow of *fMRIPrep* with smoother
(and probably more accurate) regions (e.g., for aCompCor calculations).
On the long run, we hope this will add some extra run-to-run
reproducibility to the aCompCor calculations.

Resolves: #170