FIX: Progress partial volume maps instead of posteriors (FSL FAST)
Selects the PVE maps and PVE MAP segmentation from FSL FAST to forward
it through the workflow.
PVE maps are smoother (due to the regularization of the MRF), and they
are actually used by aCompCor in its original implementation by Behzadi
(see poldracklab/fmriprep#2129).

This PR changes the contents of some derivatives and therefore targets
the next minor release.

Resolves: #212.