enh(derivatives): iterate over spaces to generate standardized conversions
Generates the anatomical derivatives in standardized volumetric spaces
directly within the derivatives workflow, instead of taking them as
inputs.

The spaces object must be passed in to determine the full specification
of output spaces.

Has some node duplicity for the generation of the reportlet, but I think
this implementation is much more clear, and follows fMRIPrep's patterns.

Resolves: poldracklab/fmriprep#2226.
Supersedes: #218.