MAINT: Mount TemplateFlow's home directory in CircleCI tests
It was a waste to cache it in the ``get_data`` job, load the appropriate cache
before ``docker run``, and finally dismiss its existence not binding the path.

This will evaluate whether the TemplateFlow home directory can be correctly
mounted from the host when using containers.

I would advocate for not overloading the docker image with pre-cached templates
and advise users to bind the appropriate path from their host (#202).

Alternative: #202.