Added function to test for FastSurfer files in --fs-subjects-dir
Checks FreeSurfer subjects dir for presence of files in mri/ with names indicating processing with FastSurfer, and returns a boolean fastsurfer_bool to indicate that FastSurfer is being used instead of Freesurfer. For development purposes, this also touches files that are expected outputs of Freesurfer, but not produced by default in FastSurfer. (Addresses #278 & https://github.com/Deep-MI/FastSurfer/issues/21)